apps/api/requirements.txt
# NeuroSync API Dependencies
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]==2.4.2
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6
httpx==0.25.0
pytest==7.4.2

# RAG System Dependencies
groq==0.4.1
openai==1.6.1
sentence-transformers==2.5.0
pinecone-client==2.2.4
numpy==1.26.4
tiktoken==0.5.1
sse-starlette==1.6.5
tenacity==8.2.3
aiohttp==3.9.1

# File Processing Dependencies
PyMuPDF==1.23.25
python-docx==1.1.0
pandas==2.2.0
chromadb==0.4.22
